Fix current code to compile
Three fixes:
- In profile controller argument for function GetUserAsync should
be taken from HttpContext (as in fix). I'm not sure, maybe there
is function like GetUserById to get them giving Id as parameter
but that exact function is taking User as an argument and if we
are thinking about current user we probably should take them from
HttpContext.

- GetUserAsync is asynchronous function and before we edit user
we need to wait for getting them back. It is why there is added
await syntax used before calling that function and that is why
function definition is different (Task<IActionResult> as return
value and async modifier.
I wonder if you shouldn't do the same with Index function as
FindByIdAsync function is also asynchronous. Not sure.

- In RegisterVM there were some annotation/decorator which doesn't
works for me. I just removed it and prayed if it will work. It
worked :) Not sure if that annotation needs to be there.

Next steps for you are to test if that works as it should.